Added Windows service parameters: name, autostart, fail-restart, log-prio
-n [--name] svcname
Allows specifying UniMRCP server service name. Thus we can run multiple UniMRCP services on a Windows machine.

-a [--autostart]
Start the service automatically after boot-up.

-f [--fail-restart] n
Restart the service automatically after n seconds if it crashed.

-l [--log-prio] priority
Set log priority for the service.

-r [--register] rootdir
This parameter has been enhanced to normalize the path and if it is relative (with respect to the current working directory), it is made absolute.