Implemented handling of TCP/MRCPv2 peer disconnect event  in server stack (Issue #162).
A new connection termination timer is set in such an event. If the control channel is closed in timely manner, then the timer is stopped. Otherwise, on expiration of the termination timeout (3 sec by default), a corresponding event is raised to the server stack, which initiates session termination and clears out all the allocated resources.

Added a new configuration parameter <termination-timeout> in <mrcpv2-uas>.