Reworked log file rotation routine.
Log file names are now generated based on current timestamp.
Not only the total number but also the age (lifetime) of log files is controlled.
All the relevant parameters are made configurable.
As a result, the function apt_log_file_open() is marked deprecated in favor of apt_log_file_open_ex().