Reworked profile-based association of MRCP resources and engines to allow specifying optional attributes per profile
All the changes are made to retain backward compatibility. Nothing has to be changed in the existing configuration or the plugin interface.

For profile-based association of resources and engines, it is advisable to use the new configuration format , which allows to specify name/value attributes, for example, as follows:

      <resource-engine-map>
        <resource id="speechsynth" engine="SS-1"/>
        <resource id="speechrecog" engine="SR-1">
          <attrib name="n1" value="v1"/>
          <attrib name="n2" value="v2"/>
        </resource>
      </resource-engine-map>

The old format is still accepted, but should be considered deprecated.

      <resource-engine-map>
        <param name="speechsynth" value="SS-1"/>
        <param name="speechrecog" value="SR-1"/>
      </resource-engine-map>