While trying to find an existing client connection, do not observe instances with closed socket. Addresses issue #259
Assume a scenario where an MRCPv2 connection is unexpectedly closed by the server. In such an event, the client stack raises an event to the application to terminate activities and release all the associated resources.

Now, what if the application does not handle such an event accordingly and attempts to use an existing connection for consecutive requests. As a result, the closed but not properly released connection could unconditionally be picked up by the client stack. The change allows to use a new connection in such a state. Note that the existing closed connection would remain in the memory until all the associated channels are closed by the application.